👷 build(aip): Add Advanced Installer project for Windows application
Introduces the Advanced Installer project file (`installer.aip`) and the application icon (`icon_256x256.ico`).

This project file is essential for generating a Windows installer for the 'Win Hotkey for iFlyVoice' application. It includes initial configurations such as product name, version (4.1.0), and links to the GitHub repository for support and updates, streamlining the distribution process

Signed-off-by: Chris K.Y. FUNG <8746768+chriskyfung@users.noreply.github.com>